Move the service into `shadowsocks.go`.
